requirements.txt
astroid==2.3.1
cycler==0.10.0
isort==4.3.21
kiwisolver==1.1.0
lazy-object-proxy==1.4.2
matplotlib==3.0.3
mccabe==0.6.1
numpy==1.17.2
pylint==2.4.2
pyparsing==2.4.2
python-dateutil==2.8.0
pyusb==1.0.2
PyVISA==1.10.1
PyVISA-py==0.3.1
six==1.12.0
typed-ast==1.4.0
wrapt==1.11.2
